Move array types
