+++ src/eckit/codec/Metadata.cc
@@ -16,7 +16,7 @@
 #include <ostream>
 #include <sstream>
 
-#include "eckit/codec/types/ArrayReference.h"
+#include "eckit/codec/types/array/ArrayReference.h"
 #include "eckit/exception/Exceptions.h"
 #include "eckit/log/JSON.h"
