+++ src/eckit/codec/codec.h
@@ -28,10 +28,7 @@
 #include "eckit/codec/detail/Reference.h"
 #include "eckit/codec/detail/StaticAssert.h"
 #include "eckit/codec/detail/sfinae.h"
-#include "eckit/codec/types/ArrayReference.h"
-#include "eckit/codec/types/StdArrayAdaptor.h"
-#include "eckit/codec/types/StdVectorAdaptor.h"
-#include "eckit/codec/types/StdVectorOfStdArrayAdaptor.h"
+#include "eckit/codec/types/array.h"
 #include "eckit/codec/types/scalar.h"
 #include "eckit/codec/types/string.h"
 #include "eckit/exception/Exceptions.h"
