+++ src/eckit/codec/types/array.h
@@ -1,0 +1,16 @@
+/*
+ * (C) Copyright 2020 ECMWF.
+ *
+ * This software is licensed under the terms of the Apache Licence Version 2.0
+ * which can be obtained at http://www.apache.org/licenses/LICENSE-2.0.
+ * In applying this licence, ECMWF does not waive the privileges and immunities
+ * granted to it by virtue of its status as an intergovernmental organisation
+ * nor does it submit to any jurisdiction.
+ */
+
+#pragma once
+
+#include "eckit/codec/types/array/ArrayReference.h"
+#include "eckit/codec/types/array/adaptors/StdArrayAdaptor.h"
+#include "eckit/codec/types/array/adaptors/StdVectorAdaptor.h"
+#include "eckit/codec/types/array/adaptors/StdVectorOfStdArrayAdaptor.h"
