+++ src/eckit/codec/types/array/ArrayMetadata.h
