+++ src/eckit/codec/types/array/ArrayReference.cc
@@ -10,7 +10,7 @@
  */
 
 
-#include "eckit/codec/types/ArrayReference.h"
+#include "eckit/codec/types/array/ArrayReference.h"
 
 #include "eckit/exception/Exceptions.h"
